replace old tone mapping with a configurable Uchimura (Gran Turismo) curve  - tried out a couple of different curves, this one looks great and has tweakables  - tone map options are set in Stage.json, and may be made animated later added lighting to particle rendering  - only makes sense for non-emissive soft particles, will need alt code for other types  - looks pretty good for smoke, dust, etc add two new stages  - Field: same meshes as TestZone for now, has a nice new skybox  - RenderDemo: not a proper stage, just a bunch of spheres for showing off PBR completely remove fake ambient light directional light colour and direction now loaded from Stage.json update mario's actions to fix size of particles, shader was accidentally squaring radius manually softened the edges of the smoke billboard textures disabled the last two mip levels of radiance cube maps to reduce artifacts flipped the z axis of the space skybox since the opposite side makes a nicer background

+++ shaders/ssao/GTAO.frag.glsl
@@ -9,7 +9,7 @@
 // NUM_OFFSETS:   number of offset iterations to do per pixel
 // LOD_BIAS:      tweak the selection of depth mipmap levels
 
-//#define FALLOFF_MODE_SIMPLE
+//#define FALLOFF_MODE_SIMPLE: don't use a thickness heuristic
 
 //============================================================================//
 
@@ -17,7 +17,7 @@
 
 layout(constant_id=0) const float INVERSE_WIDTH = 0.0;
 layout(constant_id=1) const float INVERSE_HEIGHT = 0.0;
-layout(constant_id=2) const int MAX_DEPTH_LOD = 0;
+layout(constant_id=2) const float MAX_DEPTH_LOD = 0.0;
 
 #define INVERSE_VIEWPORT vec2(INVERSE_WIDTH, INVERSE_HEIGHT)
